pyproject.toml
[tool.poetry]
name = "airpixel"
version = "0.10"
description = "Controll LEDs with arduinoNanoIOT and Python"
authors = []

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.19.1"
pyyaml = "^5.3.1"
mypy = "^0.812"
isort = "^5.7.0"

[tool.poetry.dev-dependencies]
tox = "^3.15.0"
readable-log-formatter = "^0.1.4"
pytest = "^6.2.2"
black = "^20.8b1"
flake8 = "^3.8.4"

# https://black.readthedocs.io/en/stable/compatible_configs.html
[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

